requirements.txt
Click==7.0
Flask==1.0.2
Flask-SocketIO==3.3.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
pyserial==3.4
python-engineio==3.5.1
python-socketio==3.1.2
six==1.12.0
Werkzeug==0.15.2
